update media service err handling

+++ services/media_service.go
@@ -5,8 +5,8 @@ import (
 	"database/sql"
 	"path/filepath"
 
+	"github.com/google/uuid"
 	"github.com/notefan-golang/exceptions"
-	"github.com/notefan-golang/helpers/errorh"
 	"github.com/notefan-golang/helpers/stringh"
 	"github.com/notefan-golang/models/requests/common_reqs"
 	"github.com/notefan-golang/models/requests/media_reqs"
@@ -25,19 +25,27 @@ func NewMediaService(
 }
 
 // Find finds media by the given request id
-func (service *MediaService) Find(ctx context.Context, data common_reqs.UUID) (media_ress.Media, error) {
+func (service *MediaService) Find(ctx context.Context, data common_reqs.UUID) (mediaRes media_ress.Media, err error) {
 	mediaEty, err := service.repository.Find(ctx, data.Id)
-	if err != nil { // err not nil return exception HTTPNotFound
-		errorh.Log(err)
-		return media_ress.Media{}, exceptions.HTTPNotFound
+	if err != nil {
+		return
 	}
-	return media_ress.FillFromEntity(mediaEty), nil
+	if mediaEty.Id == uuid.Nil { // if media not found return exception HTTPNotFound
+		return mediaRes, exceptions.HTTPNotFound
+	}
+	mediaRes = media_ress.FillFromEntity(mediaEty)
+	return mediaRes, nil
 }
 
 // Update updates media by the given request id
-func (service *MediaService) Update(ctx context.Context, data media_reqs.Update) (media_ress.Media, error) {
+func (service *MediaService) Update(ctx context.Context, data media_reqs.Update) (mediaRes media_ress.Media, err error) {
 	mediaEty, err := service.repository.Find(ctx, data.Id)
-	errorh.LogPanic(err) // panic if not found
+	if err != nil {
+		return
+	}
+	if mediaEty.Id == uuid.Nil { // if media not found return exception HTTPNotFound
+		return mediaRes, exceptions.HTTPNotFound
+	}
 
 	if data.Name != "" {
 		mediaEty.Name = sql.NullString{String: data.Name, Valid: true}
@@ -52,15 +60,24 @@ func (service *MediaService) Update(ctx context.Context, data media_reqs.Update)
 	}
 
 	_, err = service.repository.UpdateById(ctx, &mediaEty)
-	errorh.LogPanic(err)
+	if err != nil {
+		return
+	}
 
-	return media_ress.FillFromEntity(mediaEty), nil
+	mediaRes = media_ress.FillFromEntity(mediaEty)
+	return mediaRes, nil
 }
 
 // Delete deletes media by the given request id
-func (service *MediaService) Delete(ctx context.Context, data common_reqs.UUID) error {
+func (service *MediaService) Delete(ctx context.Context, data common_reqs.UUID) (err error) {
 	mediaEty, err := service.repository.Find(ctx, data.Id)
-	errorh.LogPanic(err) // panic if not found
+	if err != nil {
+		return
+	}
+	if mediaEty.Id == uuid.Nil { // if media not found return exception HTTPNotFound
+		err = exceptions.HTTPNotFound
+		return
+	}
 
 	_, err = service.repository.DeleteByIds(ctx, mediaEty.Id.String())
 	return err
